Add option to save visits with missing referer
The filtering of URLs can actually affect legitimate visitors too,
if their browser disables sending the `referer` despite our setting the
`referrerpolicy`. In practice, I've seen this happen only a few times,
and I'm unsure whether they were from actual people or just bots.
But even if they are real people, I think it's fine to accept their wish
to remain anonymous and just not count them anywhere.

I added this option to be able to figure out how often this happens
and in turn see how effective it is to use the referer header as a way
to get the currently visited page.